[BE] refactor: group service 리팩토링
